moved fupfms to VaultModule
